create service.rs
